refactor(entities/review): rewrite with public exports
