+++ apps/frontend/src/entities/review/api/index.ts
@@ -1,0 +1,6 @@
+export { getReviewsStats, getTestimonials } from "./review.api";
+export {
+  queryKeys as reviewQueryKeys,
+  reviewStatsOptions,
+  testimonialsOptions,
+} from "./review.queries";
