+++ apps/frontend/src/entities/review/index.ts
@@ -1,0 +1,8 @@
+export {
+  getReviewsStats,
+  getTestimonials,
+  reviewQueryKeys,
+  reviewStatsOptions,
+  testimonialsOptions,
+} from "./api";
+export { ReviewCard } from "./ui";
