+++ apps/frontend/src/entities/review/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as ReviewCard } from "./ReviewCard.vue";
